[S09P22B310-198] feat : 멤버서비스 임시구현

+++ backend/src/main/java/com/ntt/mwonimoney/domain/member/entity/Child.java
@@ -19,7 +19,7 @@ import lombok.NoArgsConstructor;
 public class Child extends Member {
 
 	@Column(name = "member_credit_score")
-	private byte creditScore;
+	private int creditScore;
 
 	@Column(name = "member_quiz_reward")
 	private int quizReward;
@@ -28,9 +28,9 @@ public class Child extends Member {
 	private SmallAccount smallAccount;
 
 	@Builder
-	public Child(byte status, String name, String nickname, String birthday,
+	public Child(int status, String name, String nickname, String birthday,
 		SocialProvider socialProvider,
-		String socialId, String email, byte creditScore, int quizReward, SmallAccount smallAccount) {
+		String socialId, String email, int creditScore, int quizReward, SmallAccount smallAccount) {
 		super(status, name, nickname, birthday, socialProvider, socialId, email, MemberRole.CHILD);
 		this.creditScore = creditScore;
 		this.quizReward = quizReward;
